Jarvis_os/requirements.txt
fastapi
uvicorn
psutil
pyautogui
pymongo
python-dotenv
python-jose
passlib[bcrypt]
email-validator